// src/app/model/local-storage.constants.ts
export enum LocalStorageConstants{
  APP_THEME = "theme",
  APP_LANGUAGE = "lang",
  Name = "name",
  Login_Status = "logInStatus",
  Auth = "auth",
  RegEmail = "regEmail",
  StartGuideStatus = "startGuideStatus",
  ProfileMediaTake = "profileMediaTake",
  ProfileMediaLibrary = "profileMediaLibrary",
  ModuleMediaTake = "moduleMediaTake",
  ModuleMediaLibrary = "moduleMediaLibrary",
  PushNotificationInstallation = "pushNotificationInstallation",
  Features = "features",
  Notification = "notification",
  NotificationValue = "notificationValue"
}